html/parsing/html5lib/html-tree/menuitem-element.dat

#data
<menuitem>
#errors
10: Start tag seen without seeing a doctype first. Expected “<!DOCTYPE html>”.
no </menuitem>
#document
| <html>
|   <head>
|   <body>
|     <menuitem>

#data
</menuitem>
#errors
11: End tag seen without seeing a doctype first. Expected “<!DOCTYPE html>”.
11: Stray end tag “menuitem”.
#document
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE html><body><menuitem>A
#errors
no </menuitem>
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <menuitem>
|       "A"

#data
<!DOCTYPE html><body><menuitem>A<menuitem>B
#errors
no </menuitem>
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <menuitem>
|       "A"
|       <menuitem>
|         "B"

#data
<!DOCTYPE html><body><menuitem>A<menu>B</menu>
#errors
no </menuitem>
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <menuitem>
|       "A"
|       <menu>
|         "B"

#data
<!DOCTYPE html><body><menuitem>A<hr>B
#errors
no </menuitem>
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <menuitem>
|       "A"
|       <hr>
|       "B"

#data
<!DOCTYPE html><li><menuitem><li>
#errors
no </menuitem>
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <li>
|       <menuitem>
|     <li>

#data
<!DOCTYPE html><menuitem><p></menuitem>x
#errors
39: Stray end tag “menuitem”.
no </menuitem>
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <menuitem>
|       <p>
|         "x"

#data
<!DOCTYPE html><p><b></p><menuitem>
#errors
21: Unclosed element “b”.
no </b> at EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <b>
|     <b>
|       <menuitem>

#data
<!DOCTYPE html><menuitem><asdf></menuitem>x
#errors
No </asdf> before </menuitem>
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <menuitem>
|       <asdf>
|     "x"

#data
<!DOCTYPE html></menuitem>
#errors
26: Stray end tag “menuitem”.
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE html><html></menuitem>
#errors
26: Stray end tag “menuitem”.
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE html><head></menuitem>
#errors
26: Stray end tag “menuitem”.
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE html><select><menuitem></select>
#errors
33: Stray start tag “menuitem”.
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>

#data
<!DOCTYPE html><option><menuitem>
#errors
no </menuitem>
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <option>
|       <menuitem>

#data
<!DOCTYPE html><menuitem><option>
#errors
no </menuitem>
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <menuitem>
|       <option>

#data
<!DOCTYPE html><menuitem></body>
#errors
no </menuitem>
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <menuitem>

#data
<!DOCTYPE html><menuitem></html>
#errors
no </menuitem>
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <menuitem>

#data
<!DOCTYPE html><menuitem><p>
#errors
no </menuitem>
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <menuitem>
|       <p>

#data
<!DOCTYPE html><menuitem><li>
#errors
no </menuitem>
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <menuitem>
|       <li>
